test: refactoring of tests binarizer structure + use fixture for reusable test data
